Add actions to handle data submission to the backend (Post)
